chore: Norm compliant and minor refactoring
